requirements.txt
    kabbes_py_starter
    kabbes_parent_class
    kabbes_dir_ops
    kabbes_menu
    kabbes_cryptography
    kabbes_nanoid
    kabbes_password_creator
    pandas
